Correct post logic
Wow I really don't like HTTParty anymore. Zero to no docs, had to read
source code to find this fix, seems like a backwards-to-http option in
the first place, and now requires me to monkey patch in by including it
and options in classes that include my module. Maybe I'm missing some
ruby feature here, but this is lame. Will want to remove this lib.